Replace Slack emoji shortcodes in event descriptions with images (#49)
Event descriptions are currently stored as HTML, so we manipulate it as a string. I think we should look into storing Markdown, considering we don't use any HTML-specific features at the moment.

We cache the emoji list and use results from previous page loads (but re-fetch in the background) because we currently need to download _every_ emoji on each page load, which can take multiple seconds on slower networks. It would be nice if Badger could only send emojis we need.